+++ tests/Benchmark/Assertion/AssertionProcessorBench.php
@@ -9,10 +9,6 @@ use PhpBench\Model\Variant;
 use PhpBench\Tests\Benchmark\IntegrationBenchCase;
 use PhpBench\Tests\Util\VariantBuilder;
 
-/**
- * @OutputTimeUnit("milliseconds")
- * @Assert("mode(variant.time.avg) as ms <= mode(baseline.time.avg) as ms +/- 5%")
- */
 class AssertionProcessorBench extends IntegrationBenchCase
 {
     /**