+++ tests/Benchmark/Expression/ParserBench.php
@@ -9,11 +9,7 @@ use PhpBench\Expression\Parser;
 use PhpBench\Extension\ExpressionExtension;
 
 /**
- * @Revs(10)
- * @Iterations(10)
  * @BeforeMethods({"setUp"})
- * @OutputTimeUnit("microseconds")
- * @Assert("mode(variant.time.avg) as ms < mode(baseline.time.avg) as ms +/- 5%")
  */
 class ParserBench
 {